+++ src/main/java/io/dutwrapper/dutwrapper/model/news/NewsGlobalGroupByDate.java
@@ -1,7 +1,6 @@
 package io.dutwrapper.dutwrapper.model.news;
 
 import java.util.ArrayList;
-import javax.annotation.Nullable;
 
 public class NewsGlobalGroupByDate {
     ArrayList<NewsGlobalItem> data;
@@ -20,7 +19,7 @@ public class NewsGlobalGroupByDate {
         return date;
     }
 
-    public void addData(NewsGlobalItem item, @Nullable Boolean addAtFirst) {
+    public void addData(NewsGlobalItem item) {
         if (item.getDate() != date) {
             return;
         }
@@ -28,35 +27,13 @@ public class NewsGlobalGroupByDate {
         if (!(data.stream().anyMatch(p -> p.getDate() == item.getDate() &&
                 p.getTitle() == item.getTitle() &&
                 p.getContent() == item.getContent()))) {
-            Boolean first = false;
-            if (addAtFirst == null)
-                first = false;
-            else if (addAtFirst == false)
-                first = false;
-
-            if (first) {
-                data.add(item);
-            } else {
-                data.add(0, item);
-            }
+            data.add(item);
         }
     }
 
-    public void addData(ArrayList<NewsGlobalItem> itemList, @Nullable Boolean addAtFirst) {
-        Boolean first = false;
-        if (addAtFirst == null)
-            first = false;
-        else if (addAtFirst == false)
-            first = false;
-
-        if (first) {
-            for (int i = itemList.size() - 1; i >= 0; i--) {
-                addData(itemList.get(i), addAtFirst);
-            }
-        } else {
-            for (NewsGlobalItem item : itemList) {
-                addData(item, addAtFirst);
-            }
+    public void addData(ArrayList<NewsGlobalItem> itemList) {
+        for (NewsGlobalItem item : itemList) {
+            addData(item);
         }
     }
 }