+++ src/main/java/io/dutwrapper/dutwrapper/model/news/NewsSubjectGroupByDate.java
@@ -1,18 +1,17 @@
 package io.dutwrapper.dutwrapper.model.news;
 
 import java.util.ArrayList;
-import javax.annotation.Nullable;
 
 public class NewsSubjectGroupByDate {
-    ArrayList<NewsGlobalItem> data;
+    ArrayList<NewsSubjectItem> data;
     long date;
 
     public NewsSubjectGroupByDate(long date) {
         this.date = date;
-        this.data = new ArrayList<NewsGlobalItem>();
+        this.data = new ArrayList<NewsSubjectItem>();
     }
 
-    public ArrayList<NewsGlobalItem> getData() {
+    public ArrayList<NewsSubjectItem> getData() {
         return data;
     }
 
@@ -20,7 +19,7 @@ public class NewsSubjectGroupByDate {
         return date;
     }
 
-    public void addData(NewsGlobalItem item, @Nullable Boolean addAtFirst) {
+    public void addData(NewsSubjectItem item) {
         if (item.getDate() != date) {
             return;
         }
@@ -28,35 +27,13 @@ public class NewsSubjectGroupByDate {
         if (!(data.stream().anyMatch(p -> p.getDate() == item.getDate() &&
                 p.getTitle() == item.getTitle() &&
                 p.getContent() == item.getContent()))) {
-            Boolean first = false;
-            if (addAtFirst == null)
-                first = false;
-            else if (addAtFirst == false)
-                first = false;
-
-            if (first) {
-                data.add(item);
-            } else {
-                data.add(0, item);
-            }
+            data.add(item);
         }
     }
 
-    public void addData(ArrayList<NewsGlobalItem> itemList, @Nullable Boolean addAtFirst) {
-        Boolean first = false;
-        if (addAtFirst == null)
-            first = false;
-        if (addAtFirst == false)
-            first = false;
-
-        if (first) {
-            for (int i = itemList.size() - 1; i >= 0; i--) {
-                addData(itemList.get(i), addAtFirst);
-            }
-        } else {
-            for (NewsGlobalItem item : itemList) {
-                addData(item, addAtFirst);
-            }
+    public void addData(ArrayList<NewsSubjectItem> itemList) {
+        for (NewsSubjectItem item : itemList) {
+            addData(item);
         }
     }
 }